Display: Windows Support
The display is now able to print on a normal console but im really not
happy with the result! I could not figure out how to print emojis
on the console on windows 8 and 10. So it will stay for the moment like
that but i will come back to the problem as soon as i have a solution!